final_system/backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
opencv-python-headless
deepface
tf-keras
librosa
scikit-learn
numpy
pillow
scipy